pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "daily-feed-digest"
version = "0.1.0"
description = "AI-powered daily digest agent for Twitter/X, GitHub, Slack, and Reddit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.25.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
digest = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
